Modified   ManaPHP/Cli/Controller.php Modified   ManaPHP/Http/Request.php Modified   ManaPHP/Model.php Modified   ManaPHP/Model/Validator.php Modified   ManaPHP/ModelInterface.php Modified   ManaPHP/Mongodb/Model.php Modified   ManaPHP/Mongodb/Query.php Modified   ManaPHP/Query.php Modified   ManaPHP/Query/Merger.php Modified   ManaPHP/QueryInterface.php

+++ ManaPHP/Cli/Controller.php
@@ -98,7 +98,7 @@ abstract class Controller extends Component implements LogCategorizable
                     } elseif ($type === 'int' || $type === 'integer') {
                         $defaultValues[$name] = (int)$defaultValues[$name];
                     } elseif ($type === 'float' || $type === 'double') {
-                        $defaultValues[$name] = (double)$defaultValues[$name];
+                        $defaultValues[$name] = (float)$defaultValues[$name];
                     } elseif ($type === 'string') {
                         $defaultValues[$name] = json_encode($defaultValues[$name], JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE);
                     } elseif ($type === 'array') {
